requirements.txt
Django
djangorestframework
pytest
pytest-django
